Warning suppression

+++ libs/kns/http-file.c
@@ -1437,7 +1437,7 @@ static rc_t KNSManagerVMakeHttpFileIntUnstableImpl( const KNSManager *self,
                                             String ext;
                                             StringInit ( & ext,
                                                     base + buf -> elem_count - vdbcache . size - 1,
-                                                    vdbcache . size, vdbcache . size );
+                                                    vdbcache . size, (uint32_t)vdbcache . size );
                                             if ( ext . addr [ ext . size ] == '\0' &&
                                                 StringEqual ( & vdbcache, & ext ) )
                                             {
@@ -1447,7 +1447,7 @@ static rc_t KNSManagerVMakeHttpFileIntUnstableImpl( const KNSManager *self,
                                                 size_t size = query - base;
                                                 StringInit ( & ext,
                                                     base + size - vdbcache . size,
-                                                    vdbcache . size, vdbcache . size );
+                                                    vdbcache . size, (uint32_t)vdbcache . size );
                                                 if ( ext . addr [ ext . size ] == '?' &&
                                                     StringEqual ( & vdbcache, & ext ) )
                                                 {
